A Simple Julia Package for SIRS Epidemiological Model

Introduction to the SIRS Model

The SIRS model is a classic epidemiological model used to describe the spread of infectious diseases within a population. It is a compartmental model that categorizes individuals into three main groups or compartments:

Susceptible (S): Individuals who are healthy but can become infected if exposed to the disease.

Infected (I): Individuals who are currently infected and can spread the disease to susceptible individuals.

Recovered (R): Individuals who have recovered from the disease and have temporary immunity.

Dynamics of the SIRS Model

In the SIRS model, individuals transition between the three compartments over time, following a specific flow:

S → I (Infection): Susceptible individuals can become infected if they come into contact with an infected person. The rate of infection depends on the contact rate and the infection probability.

I → R (Recovery): Infected individuals eventually recover from the disease, moving into the recovered category. This transition occurs at a rate determined by the average duration of the infection.

R → S (Loss of Immunity): Unlike the simpler SIR model, where recovered individuals gain permanent immunity, in the SIRS model, individuals lose immunity after a period of time and become susceptible again. This makes the SIRS model particularly useful for studying diseases where immunity is temporary, such as the flu.

Applications of the SIRS Model

The SIRS model is widely used in epidemiology to study diseases where immunity is not lifelong, such as:

Influenza (seasonal flu)

Common cold

Certain bacterial infections

By adjusting the parameters of the model (such as the infection, recovery, and immunity loss rates), the SIRS model can simulate different disease dynamics and predict the outcome of various public health interventions, such as vaccination campaigns or quarantine measures.

Limitations

While the SIRS model provides valuable insights into disease dynamics, it is still a simplified model. It assumes that all individuals are equally likely to interact, which may not be realistic for diseases that spread in structured populations or networks. Moreover, it does not account for factors such as age, mobility, or variations in immunity levels, which can also influence the spread of disease.

Nevertheless, the SIRS model serves as a foundational tool in epidemiology, allowing researchers to explore the basic principles of infectious disease modelling.

Implementation in the package

In the AgentEpidemics.jl package, the SIRS model is implemented using an agent-based approach. Rather than solving differential equations as in the classical compartmental models, this approach simulates the behavior of individual agents who transition between the Susceptible (S), Infectious (I), and Recovered (R) states. The following sections explain how different components of the SIRS model are implemented.

Agents

Each agent in the simulation is represented by an instance of the Agent struct. An agent has the following properties:

id: A unique identifier for the agent.

health: The health status of the agent, which can be 'Susceptible', 'Infected', or 'Recovered'.

location: A tuple representing the agent's position in a 2D space.

speed: The agent's speed, determining how far they can move in each time step.

contacts: A list of other agents with whom this agent has come into contact.

Here is how the Agent struct is defined:

mutable struct Agent
    id::Int
    health::String
    location::Tuple{Float64, Float64}
    speed::Float64
    contacts::Vector{Int}
end

Agent Movement

Each agent moves randomly in the simulation space during each time step. We handle this behavior by updating the location of each agent based on their speed and a random direction. The agents' movement ensures dynamic interaction between individuals, enabling them to form new contacts during the simulation, which is essential for the spread of infection.

Disease Transmission

In the SIRS model, disease transmission occurs when a susceptible agent comes into contact with an infected agent. When two agents are within the defined contact radius, and one of them is infected, the susceptible agent has a chance of becoming infected based on the infection probability.

Recovery and Immunity Loss

After a certain period, infected agents can recover based on a given recovery probability. The infected agents are then moved to the Recovered (R) state.

However, immunity in the SIRS model is temporary. After the loss of immunity, the recoved agents become susceptible again, completing the SIRS cycle.

Simulation Evolution

During each time step, the following processes occur:

Agents move within the simulation space.

Disease spreads between agents based on their contacts.

Infected agents recover and move to the recovered state.

Recovered agents lose immunity and become susceptible again.

API References

Data Structures

AgentEpidemics.Settings — Type

struct Settings

Represents the settings for the simulation.

Fields

n: The number of agents in the simulation.

total_time: The total time steps for the simulation.

initial_infection_probability: The probability that an agent is initially infected.

side_length: The size of the simulation area.

contact_radius: The maximum distance at which agents can come into contact.

mean_speed: The mean speed of the agents.

std_speed: The standard deviation of the speed of the agents.

infection_probability: The probability of infection.

recovery_probability: The probability of recovery.

immunity_loss_probability: The probability of losing immunity.

record: A boolean indicating whether to record the system state.

record_file: The name of the file to record the system state.

AgentEpidemics.SimulationOutput — Type

struct SimulationOutput

Represents the output of the simulation.

Fields

states: A dataframe representing the number of agents in each health state at each time step.

positions: A dataframe representing the position and health of each agent at each time step.

Running Simulation

AgentEpidemics.run_simulation — Function

run_simulation(n::Int, total_time::Int, initial_infection_probability::Float64, side_length::Float64, contact_radius::Float64, mean_speed::Float64, std_speed::Float64, infection_probability::Float64, recovery_probability::Float64, immunity_loss_probability::Float64, record::Bool=false, record_file::String="Timeseries.csv")

Run the simulation with the specified parameters.

Arguments

n::Int: The number of agents in the simulation.

total_time::Int: The total time steps for the simulation.

initial_infection_probability::Float64: The probability that an agent is initially infected.

side_length::Float64: The size of the simulation area.

contact_radius::Float64: The maximum distance at which agents can come into contact.

mean_speed::Float64: The mean speed of the agents.

std_speed::Float64: The standard deviation of the speed of the agents.

infection_probability::Float64: The probability of infection.

recovery_probability::Float64: The probability of recovery.

immunity_loss_probability::Float64: The probability of losing immunity.

record::Bool: A boolean indicating whether to record the system state.

record_file::String: The name of the file to record the system state.

Returns

A dataframe representing the system state at each time step.

AgentEpidemics.run_simulation — Method

run_simulation(filename::String)

Run the simulation with the settings specified in a YAML file.

Arguments

filename::String: The name of the YAML file containing the simulation settings.

Returns

A dataframe representing the system state at each time step.

AgentEpidemics.run_simulation — Method

run_simulation(settings::Settings)

Run the simulation with the specified settings.

Arguments

settings::Settings: The simulation settings.

Returns

A dataframe representing the system state at each time step.

Plotting

AgentEpidemics.plot_agents — Method

plot_agents(sim_output::SimulationOutput, time::Int)

Plots the agents in the simulation at a specific time. The agents are colored according to their health.

Arguments

sim_output::SimulationOutput: The simulation output.

time::Int: The time at which to plot the agents.

savefig::Bool=false: A boolean indicating whether to save the plot as a PNG file.

Returns

The plot object.

AgentEpidemics.plot_states — Method

plot_states(sim_output::SimulationOutput)

Plots the number of agents in each health state over time.

Arguments

sim_output::SimulationOutput: The simulation output.

savefig::Bool=false: A boolean indicating whether to save the plot as a PNG file.

Returns

The plot object.